requirements_vercel.txt
Flask==2.3.3
requests==2.31.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
ultralytics==8.0.196
Pillow==10.0.1 